[REF] shopfloor_printing_base: consolidate BaseShopfloorService
- Remove duplicate BaseShopfloorService definition in components/base.py.
- Keep the implementation in services/service.py to align with the
  base.shopfloor.service architecture.
- Fixes Pylint R8180 (consider-merging-classes-inherited) where
  multiple classes were inheriting from the same parent in one module.
