+++ shopfloor_printing_base/components/__init__.py
@@ -1,1 +1,1 @@
-from . import base, printing
+from . import printing
